Fix: 修复CBoard重构bug

+++ src/main/java/io/renren/common/utils/PageUtils.java
@@ -16,7 +16,7 @@
 
 package io.renren.common.utils;
 
-import com.baomidou.mybatisplus.plugins.Page;
+import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 
 import java.io.Serializable;
 import java.util.List;
@@ -33,11 +33,11 @@ public class PageUtils implements Serializable {
 	//总记录数
 	private int totalCount;
 	//每页记录数
-	private int pageSize;
+	private long pageSize;
 	//总页数
 	private int totalPage;
 	//当前页数
-	private int currPage;
+	private long currPage;
 	//列表数据
 	private List<?> list;
 	
@@ -75,7 +75,7 @@ public class PageUtils implements Serializable {
 		this.totalCount = totalCount;
 	}
 
-	public int getPageSize() {
+	public long getPageSize() {
 		return pageSize;
 	}
 
@@ -91,7 +91,7 @@ public class PageUtils implements Serializable {
 		this.totalPage = totalPage;
 	}
 
-	public int getCurrPage() {
+	public long getCurrPage() {
 		return currPage;
 	}
